Verify that incremental SelectNeighbors produces same result as recomputing.
Sometimes, it succeeds on a retry, after shuffling the input. And
sometimes when the result is different, the distance can also be
different. When we can find an ordering that produces the same result,
the result is correct.

However, sometimes we still hit the PANIC. That demonstrates that the
incremental algorithm indeed depends on the order being deterministic.