Usa DSA instead of one large DSM
It's more friendly to allocate the memory as we you, instead of one
large allocation. That way, we don't blow the OS caches unnecessarily,
for example.

Also, DSA has a DSA_ALLOC_NO_OOM flag which allows us to fall back
gracefully to on-disk build if enlarging the shared memory area
fails. That's broken in PostgreSQL currently though, see
https://www.postgresql.org/message-id/5efa4a5e-2b8b-42dd-80ed-f920718cf5c0%40iki.fi,
but hopefully it will get fixed.