Remove offsethash
The original motivation was to eliminate the superfluous HnswPtrAccess
call from AddToVisited. The caller has to call HnswPtrAccess() anyway,
it makes sense to pass the HnswElement rather than HnswElementPtr to
AddToVisited(). But then I realized that we can use the
pointer-variant even with DSA, because the visited-hash is
backend-private, and the addresses where the elements in shared memory
are mapped to are stable within the backend.